// Frontend/angular-app/src/app/models/model.ts
export interface Job {
  applicationDate: string|number|Date;
    appliedDate: string | number | Date;
    id: number;
    companyName: string;
    jobTitle: string;
    jobLocation: string;
    jobType: string;
    appliedOn: Date;
    salary: string;
    status: string;
    notes: string;
    ResumeUrl:string;
  }